requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.3
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3